Rename subsuper package
